Add worker code to resolc.js
